app/requirements.txt
py-cord
python-dotenv
aiohttp
datetime